Wondering where to keep DrawSurface code...
Initially I thought I should keep the DrawSurface state in the DrawController, where the state belongs. However, much of the code deals with canvas drawing, so it feels like it should be self-contained. It may be best to move most of the drawing code into a mixin, so I can reuse the tile drawing code when it comes time to create tile maps or render the game itself.