Reverting to stateful grid, might try a mixin...
It's not clear where the DrawSurface state should live. The canonical answer with Flux is to keep state in the Store, and pass props down via DrawController, but it feels really heavy and clunky. Plus, I know I'm going to need this functionality when it comes time to edit maps, so it would make sense to have a mixin that encapsulates its own state. Need to read more, and think about what I'm doing, because this is one huge component that's going to take some time to rewrite either way.