Going to rewrite DrawSurface, lots to do.
First things first, I need to decouple canvas updates from the grid state. Event handlers should write to the grid, and the actual canvas should be updated in a rAF loop, reading the grid state and updating things as intelligently as possible. That is, I should try not to redraw the whole damn thing every time. Although, that might not be a bad way to go if it's performant -- it would simplify things. Time to get on with it. Lots to consider, rewriting and restructuring DrawSurface...