Remove rAF in Keyboard. Too much latency.
Now, I need to fix the track state in the scratch track, and remove the buttons. This will probably require a new component that extends Track. I also need to fix the rendering jank when playing multiple chords, by scrolling the canvas rather than redrawing every time.

For the latter, I'll need two Pixi.Graphics objects, twice the width of the visible container. I'll have to figure out which canvas to draw to based on the time from the start (whether the duration so far, modulo msPerWidth, is even). Each new note will be drawn directly, and old notes will persist until the container is no longer visible. I'll simply adjust the position of the Graphics container to create a scrolling effect -- that should take care of the frame rate issues when chords are being played.